Make EditableTable
